Fix overflow wraparound in polynomial interleaver
The squared term of the index computation gets a bit bigger
than 2^32 with Opulent Voice parameters,  so on 32-bit platforms
the interleaver was wrapping around and corrupting 19 bytes of
the frame data.

This fix is to use 64-bit arithmetic on all platforms. The CPU
performance hit has not been evaluated.

+++ include/opvcxx/PolynomialInterleaver.h
@@ -25,7 +25,8 @@ struct PolynomialInterleaver
 
     size_t index(size_t i)
     {
-        return ((F1 * i) + (F2 * i * i)) % K;
+        return ((F1 * i) + ((uint64_t)F2 * i * i)) % K;
+        // 32-bit arithmetic is just a bit too small to handle F2*i*i for OPV.
     }
     
     void interleave(buffer_t& data)
